This is a byproduct of my work on the new OrderAwareHashMap before I
decided to fork this code to create a more customized implementation.