@Ichoran 2019-01-11
Do you have benchmark results to post?  Some of the changes are hard to gauge by eye.